Move serializers to their own module.
